changed how CoThread signales death and added getResult to Async.Await
getResult is much like apply but instead of returning the result of the promise and throwing an error if the promise is canceled, it returns a Result type which is undefined if the promise is cancled and defined with the promises result if not. It only throws an error when the promise is rejected.